Clean up excluded parser testing and create CompositeParser#isExcluded method to boost readability.
