+++ tika-core/src/main/java/org/apache/tika/parser/CompositeParser.java
@@ -70,7 +70,7 @@ public class CompositeParser extends AbstractParser {
         } else {
             this.parsers = new ArrayList<Parser>();
             for (Parser p : parsers) {
-                if (!excludeParsers.contains(p.getClass()) && !assignableFrom(excludeParsers, p.getClass())) {
+                if (!isExcluded(excludeParsers, p.getClass())) {
                     this.parsers.add(p);
                 }
             }
@@ -97,6 +97,10 @@ public class CompositeParser extends AbstractParser {
             }
         }
         return map;
+    }
+
+    private boolean isExcluded(Collection<Class<? extends Parser>> excludeParsers, Class<? extends Parser> p){
+        return excludeParsers.contains(p) || assignableFrom(excludeParsers, p);
     }
 
     private boolean assignableFrom(Collection<Class<? extends Parser>> excludeParsers, Class<? extends Parser> p) {
