--- tika-core/src/test/java/org/apache/tika/parser/DummyParserSubclass.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.tika.parser;
-
-import org.apache.tika.mime.MediaType;
-
-import java.util.Map;
-import java.util.Set;
-
-/**
- * Dummy Parser used to test Parser exclusion (TIKA-1558).
- *
- * See also {@link DummyParser} and {@link org.apache.tika.parser.mock.MockParser}.
- */
-public class DummyParserSubclass extends DummyParser {
-    public DummyParserSubclass(Set<MediaType> types, Map<String, String> metadata, String xmlText){
-        super(types, metadata, xmlText);
-    }
-}
