--- tika-core/src/test/resources/META-INF/services/org.apache.tika.parser.Parser
@@ -1,2 +1,0 @@
-org.apache.tika.parser.DummyParser
-org.apache.tika.parser.DummyParserSubclass
